feat: fixes and improvements for WooCommerce Subscriptions Gifting (#3747)
* feat(wcsg): create new class for WCSG functionality

* fix(wcsg): require shipping fields if gifted shippable product

* chore: undo unnecessary change

* fix(wcsg): new gift recipients are without password

* test: fix failing unit test
